Added parse checks that block :decompositions and :durative-actions from both being active.
The parser checks to see if the domain contains constructs that would enable constructing a durative composite action.  The semantics of this have not been developed yet (future work idea!), so for now I'm ensuring that the parser won't ever construct data structures that represent this idea (which could lead to weird behavior).