Fixed a bug in binding creation for decomposition causal links.
When specifying a causal link in a decomposition, a domain author must write the literal that the giving step contributes to the receiving step. This literal causes a binding to be added to the decomposition schema: the effect of the giving step must be bound to the precondition of the receiving step. For this binding to occur, it is necessary to verify that the literal exists as an effect/precondition of the giving/receiving step.  The bug existed in this check, because I only verified the literal existence in terms of matching predicate and arity, but I also needed to check the type of each of the terms in question.