Added code to register unexpanded composite step flaws
An unexpanded composite step flaw should be registered whenever a new step is added to the plan, and that step is composite.  A new step is only ever added when there is an open condition that must be satisfied (and a step is not being reused), so I've added code there to check whether the step is an instance of a composite action, which expands the chain of unexpanded composite step flaws when appropriate.