Added methods to check whether a given flaw selection criterion is POCL and Decompositional Complete
As mentioned in the VHPOP paper:  "it is required for completeness that every flaw matches at least one selection criterion in a flaw selection strategy." The POCL check ensures that this is the case for the FlawSelectionOrder in question. Further, in order to be Decompositionally complete, the Order must be POCL complete and must have a selection criterion for unexpanded composite steps.