Changed include directives in the decompositions header to be more conservative
It used to be the case that, because Link and LinkList were defined in the plans.h header, I had to include plans.h in decompositions.h, which caused all sorts of nightmares in terms of cyclical dependencies. Now that the respective structs have been factored out, I can be more conservative in the includes.