Expanding the functionality of the BindingList for pseudo-step swapping
In order for pseudo-steps to be made real steps, I need a way to associate a pseudo-step id that is referenced in a BindingList with a non-pseudo-step id.  Thus, I provided an auxiliary method to do so.  Because I'm accessing the Binding's internal members, I decided to make the BindingList struct a friend of the Binding struct.