Adding auxiliary static method to suggest that the compiler clean-up the given RCObject
RCObjects are used throughout the code base to track chains of steps, causal links, and flaws. Whenever a newly created chain should be cleaned up, I would find the tandem: "RCObject::ref(o); RCObject::destructive_deref(o);" which basically increments and immediately decrements (and recursively destroys) the chain in question.  Because this function is publicly available, it won't cleanup an arbitrary chain, which is why I named this function as a suggestion.