Fixed bugs around instantiating a decomposition schema.
What I originally thought was a minute change to the decomposition planning domain actually unearthed a bug regarding decomposition schema instantiation.  Turns out, I was traversing causal links in a decomposition frame's constructor in a way that could potentially duplicate certain pseudo-steps.  That has now been fixed. Also, I discovered a check during parsing that was too strict (Literal::syntactically_equal should admit terms of different types as equal as long as they are compatible. Perhaps this is not the best name for this method).